+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v1.0.1] - unreleased
+
+### Fixed
+- Updated method signatures
+
 ## [v1.0.0] - 2019-11-20
 
 ### Added
-- oe-console command: oe:module:reset-configurations
+- Reset configuration command
+
+[1.0.0]: https://github.com/OXID-eSales/developer-tools/compare/d21030a98e3cce0dd67ce3b63d74728b04592644..v1.0.0